requirements.txt
streamlit==1.30.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
requests==2.31.0
huggingface-hub==0.20.1